Stop using Gerrit instanceId as partition key
Using the Gerrit instanceId as partition key might lead to
IllegalArgument exceptions, due to the fact that it can be null (since
it is an optional configuration).

Moreover, using _only_ the instanceId would lead to a suboptimal
sharding of the messages across different kinesis shards, in case where
the majority of messages would come from one active gerrit instance.

Set the partition key to a value composed by two, always present values:
the event type and the kinesis application name.

This ensures that messages can be published regardless the instanceId
and provides a better distribution over different shards.

Bug: Issue 14697
Change-Id: I467e79d8c98958cc68911a54a684ae35afe26547