Add AWS credentials profile name parameter
Some Gerrit setups shares the AWS kinesis tables between AWS Regions,
for example multi-site setup in multiple regions. For security
reason access to the other regions should be limited to necessary
services. Add configuration param which allows to specify AWS
credentials profile to be used only by the events-aws-kinesis
plugin when connecting to the Kinesis. This profile allows us to
setup credentials or IAM Role with access to the Kinesis setup on
the other region.

Bug: Issue 16920
Change-Id: I64a1d945a113aae8c92443d9144dd3553fdb3a70