Log publishing of stream events in message_log file
Currently, when the plugin publishes a stream event to Kinesis, the
message_log file is not updated. This is in contrast with other types
of events, which are logged in the file upon successful publishing of
the messages.

Bring the logging on stream events publishing in-line with other gerrit
events. Upon successful publishing of the event, update the message_log
file with the direction, stream name and message payload.

Bug: Issue 294904654
Change-Id: I90073b82781eb3e2d19ad43f1fd89020bd233aa0