Pass stream name and consumer to KinesisConsumer
An instance of `KinesisConsumer` was constucted through a factory that
receives the stream name and the message processor as arguments.

Such arguments however were ignored and the `KinesisConsumer` was
instantiated with the corresponding class fields set as `null`.

Those class fields (`streamName` and `messageProcessor`) were then
populated when `subscribe()` was called, by providing again `streamName`
and `messageProcessor` as arguments.

This was the wrong way of populating `streamName` and
`messageProcessor`, it forced those arguments to be mutable and thus be
`null`, until the `subscribe()` method was called.

Pass stream name and consumer to KinesisConsumer via the existing
Factory by annotating them as @Assisted parameters. This allows to
remove arguments from the `subscribe()` method, since the
`KinesisConsumer` already has those set as class fields.

Change-Id: I1b102fe55b769f4b2899d4ff3fa7e64d36355fa6